% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_discover.R
\name{dna_discover}
\alias{dna_discover}
\title{Run predictor for presence of the platform bias}
\usage{
dna_discover(data, P = 1e-05, auc = TRUE)
}
\arguments{
\item{data}{data.frame with each column representing specific variant
annotation type. Annotations that could be included:
group, lcr,decoy, segdup,nonpar, variant_type, allele_type, was_mixed,
has_star,qd,info_MQRankSum, info_SOR, info_InbreedingCoeff, info_FS,
info_QD, info_MQ, info_DP, rf_probability, was_split, score,
qual, BaseQRankSum, ClippingRankSum,
FS, InbreedingCoeff, MQ, MQRankSum, QD, ReadPosRankSum.
For specific annotation descriptions please look up gnomad.broadinstitute.org}

\item{P}{a threshold for significance of discordance that was observed in
gnomad (exomes vs genomes in NFE population). Default value is 1e-5.
Smaller values increase specificity, greater values increase sensitivity.}

\item{auc}{boolean. If the ROC AUC estimate on training data should be printed}
}
\value{
boolean vector of predictions
}
\description{
Run predictor for presence of the platform bias
}
