% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Training dataset with variant annotations from 100 random variants from
gnomAD genomes dataset}
\format{
A data frame with 100 rows and 65 columns
}
\usage{
test_data
}
\description{
Training dataset with variant annotations from 100 random variants from
gnomAD genomes dataset
}
\keyword{datasets}
